Added the `Headers` struct. This struct is a wrapper over a StringHashMap used to store HTTP header information.
The difference between `Headers` and a regular hash map is, in general,
the following:

1. The header values are interned in a single data buffer. The backing
hash map doesn't store raw strings, but instead start and end indicies
referencing the buffer.
2. Keys are duplicated in a separate buffer. This is because I had
issues iterating over the keys - there tended to be extra keys that
appear to be variations on similar keys added sequentially.
3. Header names are coerced to lowercase before being stored. This is to
make it easier to find them later. Header values are not changed.